Commiting changes by GrayGnome
ersion 1.2 Sat Dec  3 Sat Dec  3 09:15:52 CST 2011
Modified Thu Sep  8 07:33:17 CDT 2011 by GreyGnome.  Fixes a bug with the initial port value.  Now it sets the initial value to be the state of the port at the time of attachInterrupt().  The line is port.PCintLast=port.portInputReg; in attachInterrupt().

Added the "arduinoPin" variable, so the user's function will know exactly which pin on the Arduino was triggered.  See the code.

Finally, changed the install procedure:  To install, just unzip in your Arduino libraries folder.  The zipfile contains the subdirectories and files, so it should install it properly for use by Arduino.

To find the Arduino libraries folder, open up your Arduino IDE preferences and look at the Sketchbook location.  Your libraries are simply in the "libraries" subfolder there.

You must exit the IDE and reenter in order to see the libraries.  Go to the Sketch menu, and look at the Import Library submenu.  You should see the AdaEncoder and PinChangeInt libraries listed under "Contributed".


-Lex